requirements.txt
fastapi
uvicorn
pandas
plotly
streamlit
scikit-learn
requests
seaborn
matplotlib
google-generativeai
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-logging
google-cloud-storage
econml
gTTS
scipy
streamlit-autorefresh
python-dotenv
supabase
